change the `multiple_of` of the Schema from integer to float, to support the situation: `multiple_of` will be `0.01`, when the numeric need to hold two decimals, and any greater precision will be rejected.

+++ java/src/main/java/org/mojolang/mojo/openapi/SchemaOrBuilder.java
@@ -118,10 +118,10 @@ public interface SchemaOrBuilder extends
       getFormatBytes();
 
   /**
-   * <code>int64 multiple_of = 14;</code>
+   * <code>double multiple_of = 14;</code>
    * @return The multipleOf.
    */
-  long getMultipleOf();
+  double getMultipleOf();
 
   /**
    * <code>double maximum = 15;</code>
